+++ src/hardware/mod.rs
@@ -4,6 +4,7 @@ pub use stm32h7xx_hal as hal;
 
 pub mod adc;
 pub mod afe;
+pub mod cpu_temp_sensor;
 pub mod dac;
 pub mod design_parameters;
 pub mod input_stamper;
@@ -57,6 +58,10 @@ pub type EthernetPhy = hal::ethernet::phy::LAN8742A<hal::ethernet::EthernetMAC>;
 pub const MONOTONIC_FREQUENCY: u32 = 1_000;
 pub type Systick = systick_monotonic::Systick<MONOTONIC_FREQUENCY>;
 pub type SystemTimer = mono_clock::MonoClock<u32, MONOTONIC_FREQUENCY>;
+
+pub type I2c1 = hal::i2c::I2c<hal::stm32::I2C1>;
+pub type I2c1Proxy =
+    shared_bus::I2cProxy<'static, shared_bus::AtomicCheckMutex<I2c1>>;
 
 #[inline(never)]
 #[panic_handler]